fix: have `/health` and `/info` endpoints
