+++ src/lib.rs
@@ -197,6 +197,7 @@ pub async fn bootstap(mut shutdown: broadcast::Receiver<()>, config: Config) -> 
 
         Router::new()
             .route("/health", get(handlers::health::handler))
+            .route("/info", get(handlers::info::handler))
             .nest("/tenants", tenancy_routes)
             .route(
                 "/:tenant_id/clients",
@@ -217,6 +218,7 @@ pub async fn bootstap(mut shutdown: broadcast::Receiver<()>, config: Config) -> 
     #[cfg(not(feature = "multitenant"))]
     let app = Router::new()
         .route("/health", get(handlers::health::handler))
+        .route("/info", get(handlers::info::handler))
         .route(
             "/clients",
             post(handlers::single_tenant_wrappers::register_handler),
